TKSS-442: SM2 cipher should accept empty input
